extracted month picker to its own file
